Add Palemoon support
A new task has been introduced to gulp that allows building an extension
compatible with the Palemoon browser. Palemoon extensions need to be
built using Mozilla's old CFX tool, which unfortunately requires Python
2.7 to be installed. This is also the main reason why `build:palemoon`
is not part of the `build` task.
The source to built from is the Firefox extension, since there are no
incompatibilities in the Firefox code we use, it was just the build
tool being incompatible